remove User and merge fields into RoomMember
