+++ src/models/mod.rs
@@ -1,10 +1,8 @@
 mod room;
 mod room_member;
-mod user;
 
 pub use room::{Room, RoomName};
 pub use room_member::RoomMember;
-pub use user::User;
 
 #[allow(dead_code)]
 pub type Token = String;
