+++ src/models/room_member.rs
@@ -15,15 +15,15 @@
 
 use std::convert::TryFrom;
 
-use super::User;
 use crate::events::collections::all::Event;
 use crate::events::room::{
     member::{MemberEvent, MembershipChange, MembershipState},
     power_levels::PowerLevelsEvent,
 };
+use crate::events::presence::{PresenceEvent, PresenceEventContent, PresenceState};
 use crate::identifiers::UserId;
 
-use js_int::Int;
+use js_int::{Int, UInt};
 
 // Notes: if Alice invites Bob into a room we will get an event with the sender as Alice and the state key as Bob.
 
@@ -33,12 +33,22 @@ use js_int::Int;
 pub struct RoomMember {
     /// The unique mxid of the user.
     pub user_id: UserId,
+    /// The human readable name of the user.
+    pub display_name: Option<String>,
+    /// The matrix url of the users avatar.
+    pub avatar_url: Option<String>,
+    /// The time, in ms, since the user interacted with the server.
+    pub last_active_ago: Option<UInt>,
+    /// If the user should be considered active.
+    pub currently_active: Option<bool>,
     /// The unique id of the room.
     pub room_id: Option<String>,
     /// If the member is typing.
     pub typing: Option<bool>,
-    /// The user data for this room member.
-    pub user: User,
+    /// The presence of the user, if found.
+    pub presence: Option<PresenceState>,
+    /// The presence status message, if found.
+    pub status_msg: Option<String>,
     /// The users power level.
     pub power_level: Option<Int>,
     /// The normalized power level of this `RoomMember` (0-100).
@@ -49,20 +59,27 @@ pub struct RoomMember {
     pub name: String,
     /// The events that created the state of this room member.
     pub events: Vec<Event>,
+    /// The `PresenceEvent`s connected to this user.
+    pub presence_events: Vec<PresenceEvent>,
 }
 
 impl RoomMember {
     pub fn new(event: &MemberEvent) -> Self {
-        let user = User::new(event);
         Self {
+            name: event.state_key.clone(),
             room_id: event.room_id.as_ref().map(|id| id.to_string()),
             user_id: UserId::try_from(event.state_key.as_str()).unwrap(),
+            display_name: event.content.displayname.clone(),
+            avatar_url: event.content.avatar_url.clone(),
+            presence: None,
+            status_msg: None,
+            last_active_ago: None,
+            currently_active: None,
             typing: None,
-            user,
             power_level: None,
             power_level_norm: None,
             membership: event.content.membership,
-            name: event.state_key.clone(),
+            presence_events: Vec::default(),
             events: vec![Event::RoomMember(event.clone())],
         }
     }
@@ -72,8 +89,8 @@ impl RoomMember {
 
         match event.membership_change() {
             ProfileChanged => {
-                self.user.display_name = event.content.displayname.clone();
-                self.user.avatar_url = event.content.avatar_url.clone();
+                self.display_name = event.content.displayname.clone();
+                self.avatar_url = event.content.avatar_url.clone();
                 true
             }
             Banned | Kicked | KickedAndBanned | InvitationRejected | InvitationRevoked | Left
@@ -109,6 +126,64 @@ impl RoomMember {
         }
 
         changed
+    }
+
+    /// If the current `PresenceEvent` updated the state of this `User`.
+    ///
+    /// Returns true if the specific users presence has changed, false otherwise.
+    ///
+    /// # Arguments
+    ///
+    /// * `presence` - The presence event for a this room member.
+    pub fn did_update_presence(&self, presence: &PresenceEvent) -> bool {
+        let PresenceEvent {
+            content:
+                PresenceEventContent {
+                    avatar_url,
+                    currently_active,
+                    displayname,
+                    last_active_ago,
+                    presence,
+                    status_msg,
+                },
+            ..
+        } = presence;
+        self.display_name == *displayname
+            && self.avatar_url == *avatar_url
+            && self.presence.as_ref() == Some(presence)
+            && self.status_msg == *status_msg
+            && self.last_active_ago == *last_active_ago
+            && self.currently_active == *currently_active
+    }
+
+    /// Updates the `User`s presence.
+    ///
+    /// This should only be used if `did_update_presence` was true.
+    ///
+    /// # Arguments
+    ///
+    /// * `presence` - The presence event for a this room member.
+    pub fn update_presence(&mut self, presence_ev: &PresenceEvent) {
+        let PresenceEvent {
+            content:
+                PresenceEventContent {
+                    avatar_url,
+                    currently_active,
+                    displayname,
+                    last_active_ago,
+                    presence,
+                    status_msg,
+                },
+            ..
+        } = presence_ev;
+
+        self.presence_events.push(presence_ev.clone());
+        self.avatar_url = avatar_url.clone();
+        self.currently_active = currently_active.clone();
+        self.display_name = displayname.clone();
+        self.last_active_ago = last_active_ago.clone();
+        self.presence = Some(presence.clone());
+        self.status_msg = status_msg.clone();
     }
 }
